requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
pymongo==4.14.1
python-dotenv==1.1.1
google-api-python-client==2.179.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
dateparser==1.2.2
httpx==0.28.1
python-dateutil==2.9.0.post0
protobuf==4.25.8
SpeechRecognition==3.10.4
python-multipart==0.0.20
